# LMS/docs/swagger.yaml
basePath: /lmsapiv1
definitions:
  dto.ContentResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      id:
        type: integer
      is_mandatory:
        type: boolean
      is_published:
        type: boolean
      metadata:
        additionalProperties: true
        type: object
      order_index:
        type: integer
      section_id:
        type: integer
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.UpdateContentRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      is_mandatory:
        type: boolean
      is_published:
        type: boolean
      metadata:
        additionalProperties: true
        type: object
      order_index:
        minimum: 0
        type: integer
      title:
        maxLength: 255
        minLength: 3
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@bdc.hpcc.vn
    name: API Support
    url: https://bdc.hpcc.vn/support
  description: This is the API documentation for LMS (Learning Management System)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://bdc.hpcc.vn/terms
  title: LMS API Documentation
  version: "1.0"
paths:
...
  /content/{contentId}:
    delete:
      consumes:
      - application/json
      description: Delete content (Owner/Admin only)
      parameters:
      - description: Content ID
        in: path
        name: contentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete content
      tags:
      - content
    get:
      consumes:
      - application/json
      description: Get content details by ID
      parameters:
      - description: Content ID
        in: path
        name: contentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ContentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get content
      tags:
      - content
    put:
      consumes:
      - application/json
      description: Update content details (Owner/Admin only)
      parameters:
      - description: Content ID
        in: path
        name: contentId
        required: true
        type: integer
      - description: Content data
        in: body
        name: content
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateContentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update content
      tags:
      - content
  /courses:
    get:
      consumes:
      - application/json
      description: List all published courses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CourseResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List published courses
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Create a new course (Teacher/Admin only)
      parameters:
      - description: Course data
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CourseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a course
      tags:
      - courses
  /courses/{courseId}:
    delete:
      consumes:
      - application/json
      description: Delete a course (Admin only)
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a course
      tags:
      - courses
    get:
      consumes:
      - application/json
      description: Get course details by ID
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CourseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a course
      tags:
      - courses
    put:
      consumes:
      - application/json
      description: Update course details (Owner/Admin only)
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: integer
      - description: Course data
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a course
      tags:
      - courses
  /courses/{courseId}/publish:
    post:
      consumes:
      - application/json
      description: Publish a course (Owner/Admin only)
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Publish a course
      tags:
      - courses
  /courses/{courseId}/sections:
    get:
      consumes:
      - application/json
      description: List all sections in a course
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SectionResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List sections
      tags:
      - sections
    post:
      consumes:
      - application/json
      description: Create a new section in a course (Owner/Admin only)
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: integer
      - description: Section data
        in: body
        name: section
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SectionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a section
      tags:
      - sections
  /courses/my:
    get:
      consumes:
      - application/json
      description: List all courses created by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CourseResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my courses
      tags:
      - courses
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"